+++ tests/test_optimization.py
@@ -6,7 +6,6 @@ import numpy as np
 
 from glassure import Pattern, convert_density_to_atoms_per_cubic_angstrom
 from glassure.utility import (
-    extrapolate_to_zero_poly,
     extrapolate_to_zero_linear,
     calculate_f_mean_squared,
     calculate_f_squared_mean,
@@ -17,7 +16,7 @@ from glassure.transform import calculate_sq, calculate_fr
 from glassure.normalization import normalize_fit, normalize
 from glassure.configuration import OptimizeConfig
 from glassure.optimization import optimize_sq, optimize_density
-from glassure.calc import calculate_pdf, create_calculate_pdf_configs
+from glassure.calc import create_calculate_pdf_configs
 from glassure.methods import ExtrapolationMethod
 
 from . import unittest_data_path
@@ -115,15 +114,24 @@ def test_optimize_sq(sq, atomic_density):
 
 def test_optimize_sq_fft(sq, atomic_density):
     iterations = 5
-    r_step = 0.001 # need high value to be accurate for fft
+    r_step = 0.001  # need high value to be accurate for fft
     sq_optimized = optimize_sq(
-        sq, 1.3, iterations, atomic_density, fourier_transform_method="integral", r_step=r_step
+        sq,
+        1.3,
+        iterations,
+        atomic_density,
+        fourier_transform_method="integral",
+        r_step=r_step,
     )
     fr_optimized = calculate_fr(sq_optimized, method="fft")
 
-
     sq_optimized_fft = optimize_sq(
-        sq, 1.3, iterations, atomic_density, fourier_transform_method="fft", r_step=r_step
+        sq,
+        1.3,
+        iterations,
+        atomic_density,
+        fourier_transform_method="fft",
+        r_step=r_step,
     )
     fr_optimized_fft = calculate_fr(sq_optimized_fft, method="fft")
 
@@ -131,7 +139,34 @@ def test_optimize_sq_fft(sq, atomic_density):
     assert np.allclose(fr_optimized.y, fr_optimized_fft.y, atol=0.1)
 
 
-def test_optimize_density():
+def test_optimize_density_SiO2(data_path, bkg_path):
+    data = Pattern.from_file(data_path)
+    background = Pattern.from_file(bkg_path)
+    composition = {"Si": 1, "O": 2}
+    density = 2.2
+
+    data_config, calculation_config = create_calculate_pdf_configs(
+        data, composition, density, background
+    )
+
+    calculation_config.transform.q_min = 1
+    calculation_config.transform.q_max = 16
+    calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
+    calculation_config.optimize = OptimizeConfig(r_cutoff=1.4)
+    calculation_config.transform.fourier_transform_method = "fft"
+
+    density, density_error, bkg_scaling, bkg_scaling_error = optimize_density(
+        data_config, calculation_config, method="fr"
+    )
+
+    assert density > 0
+    assert density != 2.2
+    assert density_error > 0
+    assert bkg_scaling != 1.0
+    assert bkg_scaling_error > 0
+
+
+def test_optimize_density_Mg2SiO4():
     data = Pattern.from_file(data_path_glass)
     background = Pattern.from_file(background_path_glass)
     composition = {"Mg": 2, "Si": 1, "O": 4}
@@ -145,41 +180,19 @@ def test_optimize_density():
     calculation_config.transform.q_max = 16
     calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
     calculation_config.optimize = OptimizeConfig(r_cutoff=1.4)
+    calculation_config.transform.fourier_transform_method = "fft"
 
-    density, density_error = optimize_density(data_config, calculation_config)
-
+    density, density_error, bkg_scaling, bkg_scaling_error = optimize_density(
+        data_config, calculation_config, method="fr"
+    )
     assert density > 0
     assert density != 3.21
     assert density_error > 0
+    assert bkg_scaling != 1.0
+    assert bkg_scaling_error > 0
 
 
-def test_optimize_density_x_range():
-    data = Pattern.from_file(data_path_glass)
-    background = Pattern.from_file(background_path_glass)
-    composition = {"Mg": 2, "Si": 1, "O": 4}
-    density = 3.21
-
-    data_config, calculation_config = create_calculate_pdf_configs(
-        data, composition, density, background
-    )
-
-    calculation_config.transform.q_min = 1
-    calculation_config.transform.q_max = 16
-    calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
-    calculation_config.optimize = OptimizeConfig(r_cutoff=1.2)
-
-    density_1, density_error_1 = optimize_density(
-        data_config, calculation_config, min_range=(0.3, 1.4)
-    )
-
-    density_2, density_error_2 = optimize_density(
-        data_config, calculation_config, min_range=(0.3, 0.7)
-    )
-
-    assert density_1 != density_2
-
-
-def test_optimize_density_type():
+def test_optimize_density_method():
     data = Pattern.from_file(data_path_glass)
     background = Pattern.from_file(background_path_glass)
     composition = {"Mg": 2, "Si": 1, "O": 4}
@@ -197,16 +210,16 @@ def test_optimize_density_type():
     density_gr, density_gr_error = optimize_density(
         data_config,
         calculation_config,
-        type="gr",
+        method="gr",
         min_range=(0.3, 1.0),
-        method="lsq",
+        optimization_method="lsq",
     )
 
     density_sq, density_sq_error = optimize_density(
         data_config,
         calculation_config,
-        type="sq",
-        method="lsq",
+        method="sq",
+        optimization_method="lsq",
     )
 
     assert density_gr != density_sq
@@ -228,12 +241,12 @@ def test_optimize_density_method():
     calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
     calculation_config.optimize = OptimizeConfig(r_cutoff=1.4)
 
-    density_gr, density_gr_residual = optimize_density(
-        data_config, calculation_config, type="sq", method="nelder"
+    density_gr, density_gr_residual, _, _ = optimize_density(
+        data_config, calculation_config, method="sq", optimization_method="nelder"
     )
 
-    density_sq, density_sq_residual = optimize_density(
-        data_config, calculation_config, type="sq", method="lsq"
+    density_sq, density_sq_residual, _, _ = optimize_density(
+        data_config, calculation_config, method="sq", optimization_method="lsq"
     )
 
     assert density_gr_residual != density_sq_residual
@@ -243,17 +256,63 @@ def test_optimize_density_method():
     import time
 
     start_time = time.time()
-    density_gr, density_gr_residual = optimize_density(
-        data_config, calculation_config, type="sq", method="nelder"
+    density_gr, density_gr_residual, _, _ = optimize_density(
+        data_config, calculation_config, method="sq", optimization_method="nelder"
     )
     nelder_time = time.time() - start_time
 
     start_time = time.time()
-    density_sq, density_sq_residual = optimize_density(
-        data_config, calculation_config, type="sq", method="lsq"
+    density_sq, density_sq_residual, _, _ = optimize_density(
+        data_config, calculation_config, method="sq", optimization_method="lsq"
     )
     least_squares_time = time.time() - start_time
 
-    print(f"Nelder time: {nelder_time}, Least squares time: {least_squares_time}")
     assert nelder_time > least_squares_time
-    print(f"Density: {density_gr}, {density_sq}")
+
+
+def test_optimize_density_vary_bkg_scaling():
+    data = Pattern.from_file(data_path_glass)
+    background = Pattern.from_file(background_path_glass)
+    composition = {"Mg": 2, "Si": 1, "O": 4}
+    density = 3.21
+
+    data_config, calculation_config = create_calculate_pdf_configs(
+        data, composition, density, background
+    )
+
+    calculation_config.transform.q_min = 1
+    calculation_config.transform.q_max = 16
+    calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
+    calculation_config.optimize = OptimizeConfig(r_cutoff=1.4)
+
+    density_1, density_error, bkg_scaling, bkg_scaling_error = optimize_density(
+        data_config, calculation_config, method="fr", vary_bkg_scaling=False
+    )
+    assert bkg_scaling == 1.0
+    assert bkg_scaling_error == 0
+
+    density_2, _, bkg_scaling, bkg_scaling_error = optimize_density(
+        data_config, calculation_config, method="fr", vary_bkg_scaling=True
+    )
+    assert bkg_scaling != 1.0
+    assert bkg_scaling_error > 0
+    assert density_1 != density_2
+
+
+def test_invalid_optimize_density_method():
+    data = Pattern.from_file(data_path_glass)
+    background = Pattern.from_file(background_path_glass)
+    composition = {"Mg": 2, "Si": 1, "O": 4}
+    density = 3.21
+
+    data_config, calculation_config = create_calculate_pdf_configs(
+        data, composition, density, background
+    )
+
+    with pytest.raises(ValueError):
+        optimize_density(data_config, calculation_config, method="gr", min_range=None)
+
+    with pytest.raises(ValueError):
+        optimize_density(data_config, calculation_config, method="fr", min_range=None)
+
+    res = optimize_density(data_config, calculation_config, method="sq", min_range=None)