+++ changelog.md
@@ -3,6 +3,8 @@
 ### New features
 
 - added xraylib scattering factor calculator, which can be used as an alternative to the Hajdu and Brown-Hubbell calculators - choose it by setting sf_source to 'xraylib' in the corresponding functions or in the configuration
+- optimize_density was reworked and extended. It now supports optimizing the density and background scaling at the same time. The "fr" method was added
+  and is the default method, since it the method described in Eggert et al. 2002 and other similar following publications.
 
 ### Bugfixes